simplify sound.rs implementation
